+++ mrfilestorage/wrapper_errors.go
@@ -12,14 +12,14 @@ import (
 
 func (fp *FileProvider) wrapError(err error) error {
 	if errors.Is(err, os.ErrNotExist) {
-		return mrcore.FactoryErrStorageNoRowFound.Caller(1).Wrap(err)
+		return mrcore.FactoryErrStorageNoRowFound.Wrap(err)
 	}
 
 	if _, ok := err.(*fs.PathError); ok {
-		return mrcore.FactoryErrStorageQueryFailed.Caller(1).Wrap(err)
+		return mrcore.FactoryErrStorageQueryFailed.WithCaller(1).Wrap(err)
 	}
 
-	return mrcore.FactoryErrInternal.Caller(1).Wrap(err)
+	return mrcore.FactoryErrInternal.WithCaller(1).Wrap(err)
 }
 
 func (fp *FileProvider) traceCmd(ctx context.Context, command, filePath string) {