+++ mrpostgres/conn_transaction.go
@@ -10,7 +10,7 @@ func (c *ConnAdapter) Begin(ctx context.Context) (mrstorage.DBTransaction, error
 	tx, err := c.pool.Begin(ctx)
 
 	if err != nil {
-		return nil, wrapError(err, skipThisMethod)
+		return nil, wrapError(err, skipThisMethodFrame)
 	}
 
 	return &Transaction{tx: tx}, nil