+++ mrpostgres/wrapper_errors.go
@@ -12,22 +12,22 @@ import (
 )
 
 const (
-	skipThisMethod = 1
+	skipThisMethodFrame = 1
 )
 
-func wrapError(err error, skip int) error {
+func wrapError(err error, skipFrame int) error {
 	_, ok := err.(*pgconn.PgError)
 
 	if ok {
 		// Severity: ERROR; Code: 42601; Message syntax error at or near "item_status"
-		return mrcore.FactoryErrStorageQueryFailed.Caller(skip + 1).Wrap(err)
+		return mrcore.FactoryErrStorageQueryFailed.WithCaller(skipFrame + 1).Wrap(err)
 	}
 
 	if errors.Is(err, pgx.ErrNoRows) {
-		return mrcore.FactoryErrStorageNoRowFound.Caller(skip + 1).Wrap(err)
+		return mrcore.FactoryErrStorageNoRowFound.Wrap(err)
 	}
 
-	return mrcore.FactoryErrInternal.Caller(skip + 1).Wrap(err)
+	return mrcore.FactoryErrInternal.WithCaller(skipFrame + 1).Wrap(err)
 }
 
 func traceQuery(ctx context.Context, sql string) {