+++ mrrabbitmq/wrapper_errors.go
@@ -3,5 +3,5 @@ package mrrabbitmq
 import "github.com/mondegor/go-webcore/mrcore"
 
 func (c *ConnAdapter) wrapError(err error) error {
-	return mrcore.FactoryErrStorageQueryFailed.Caller(1).Wrap(err)
+	return mrcore.FactoryErrStorageQueryFailed.WithCaller(1).Wrap(err)
 }